requirements.txt
flask
requests
fpdf2
pandas
